[REFACTOR][WL] move context providers to its own folder
