+++ src/components/elements/wrapper/types.ts
@@ -1,7 +1,7 @@
 import { IFilterCheckboxSchema } from "../../custom/filter/filter-checkbox/types";
 import { IFilterItemSchema } from "../../custom/filter/filter-item/types";
 import { IColumns, TComponentOmitProps, TFrontendEngineFieldSchema } from "../../frontend-engine";
-import { TRenderRules } from "../../frontend-engine/yup";
+import { TRenderRules } from "../../../context-providers";
 
 export type TWrapperType = "div" | "span" | "header" | "footer" | "h1" | "h2" | "h3" | "h4" | "h5" | "h6" | "p";
