+++ src/components/fields/file-upload/types.ts
@@ -1,5 +1,5 @@
 import { FileItemProps, FileUploadProps } from "@lifesg/react-design-system";
-import { IYupValidationRule } from "../../frontend-engine/yup/types";
+import { IYupValidationRule } from "../../..//context-providers";
 import { IBaseFieldSchema } from "../types";
 
 export type TUploadType = "base64" | "multipart";
